move modals to separate components, clean up
