+++ src/components/Modals/FlagModal.jsx
@@ -1,0 +1,41 @@
+import React, { useState } from 'react';
+import styles from './Modals.module.css';
+
+export const FlagModal = ({ isOpen, onClose, onSubmit }) => {
+    const [reason, setReason] = useState('');
+
+    const handleSubmit = (e) => {
+        e.preventDefault();
+        onSubmit(reason);
+        setReason('');
+    };
+
+    if (!isOpen) return null;
+
+    return (
+        <div className={styles.modalOverlay}>
+            <div className={styles.modalContent}>
+                <h3>Flag Quote</h3>
+                <p>
+                    Please provide a reason for flagging this quote:
+                </p>
+                <form onSubmit={handleSubmit}>
+                    <textarea
+                        value={reason}
+                        onChange={(e) => setReason(e.target.value)}
+                        placeholder="Enter your reason here..."
+                        required
+                    />
+                    <div className={styles.modalButtons}>
+                        <button type="button" onClick={onClose}>
+                            Cancel
+                        </button>
+                        <button type="submit">
+                            Submit
+                        </button>
+                    </div>
+                </form>
+            </div>
+        </div>
+    );
+};
